Move EnvEffects class to MetaAudio namespace
Also moved it to other directory
